+++ README.md
@@ -1,6 +1,6 @@
 # skrebets-sprint11.github.io
 
-*version 0.0.6*
+*version 0.0.7*
 
 ## Link
