+++ src/scripts/pop-up-edit.js
@@ -1,14 +1,20 @@
 /* Класс попапа формы «Редактировать профиль» */
-
-import {fillInput} from '../index.js';
+import {formEdit} from './validation.js';
 
 export const errorUser = document.querySelector('.error-user');
 export const errorJob = document.querySelector('.error-job');
 export const profileEditContainer = document.querySelector('.profile-edit');
+export const userName = document.querySelector('.user-info__name');
+export const userJob = document.querySelector('.user-info__job');
 
 export class PopUpEdit {
     constructor(container) {
       this.container = container;
+    }
+
+    fillInput() {
+      formEdit.elements.user.value = userName.textContent;
+      formEdit.elements.job.value = userJob.textContent;
     }
   
     open() {
@@ -19,7 +25,9 @@ export class PopUpEdit {
   
       errorJob.textContent = '';
   
-      fillInput();
+      const popUpEdit = new PopUpEdit(profileEditContainer);
+
+      popUpEdit.fillInput();
     }
   
     close() {
